python/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "openpowerlifting"
version = "0.1.1"
authors = [{name = "Anthony Moszuti", email = "tonymoszuti@users.noreply.github.com"}]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
requires-python=">=3.7"

dependencies = [
    "dynaconf==3.2.6",
    "langchain==0.2.12",
    "langchain-core==0.2.28",
    "langchain-community>=0.2.10",
    "langchain-openai==0.1.20",
    "openai==1.37.1",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.1",
    "sqlalchemy>=2.0.31"
]

[project.optional-dependencies]
dev = [
    "openpowerlifting[test]"
]

[project.urls]
Home = "https://github.com/tonymoszuti/openpowerlifting"